change file directory

+++ src/main/java/com/epam/java_basic/calculator/Calculator.java
@@ -7,19 +7,19 @@ public class Calculator {
     }
 
     public double add(double a, double b) {
-        int temp = a + b;
+        double temp = a + b;
         return (double) Math.round(temp * 100) / 100;
         // throw new UnsupportedOperationException("You need to implement this method");
     }
 
     public double subtract(double a, double b) {
-        int temp = a - b;
+        double temp = a - b;
         return (double) Math.round(temp * 100) / 100;
         // throw new UnsupportedOperationException("You need to implement this method");
     }
 
     public double multiply(double a, double b) {
-        int temp = a * b;
+        double temp = a * b;
         return (double) Math.round(temp * 100) / 100;
         // throw new UnsupportedOperationException("You need to implement this method");
     }
